+++ CHANGELOG.md
@@ -1,4 +1,15 @@
+## v0.2.0
+
+beta release
+
+i consider this to be near functional completion at least
+
+- add a global, script-defined callback for tuning changes
+- `add_tuning` API function
+
+
 ## v0.1.0
+
 
 2nd alpha release